extract Checkboxes
