+++ platform/jewel/samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/ComponentsViewModel.kt
@@ -10,7 +10,7 @@ import org.jetbrains.jewel.samples.showcase.components.Banners
 import org.jetbrains.jewel.samples.standalone.StandaloneSampleIcons
 import org.jetbrains.jewel.samples.showcase.components.Borders
 import org.jetbrains.jewel.samples.showcase.components.Buttons
-import org.jetbrains.jewel.samples.standalone.view.component.Checkboxes
+import org.jetbrains.jewel.samples.showcase.components.Checkboxes
 import org.jetbrains.jewel.samples.standalone.view.component.ChipsAndTrees
 import org.jetbrains.jewel.samples.standalone.view.component.Dropdowns
 import org.jetbrains.jewel.samples.standalone.view.component.Icons
